debugs, colors, release changes
