+++ app/appdata/setups/GUI.json
@@ -1,1 +1,1 @@
-{"Assistant": "pan", "UserInterface": "flask_local", "Lexicon": "classic", "Formal": "analytic_freedom", "Output": "actual_tree"}
+{"Assistant": "pan", "UserInterface": "flask_local", "Lexicon": "classic", "Formal": "analytic_signed", "Output": "actual_tree"}
