+++ app/core/colors.py
@@ -1,7 +1,3 @@
-"""
-Based on https://sashamaps.net/docs/resources/20-colors/
-Animal names from https://list.fandom.com/wiki/List_of_common_animals
-"""
 from dataclasses import dataclass
 from typing import Iterable
 from random import Random
@@ -28,177 +24,28 @@ def get_branch_name(accessibility: int, used: Iterable[str]) -> Iterable[str]:
             yield i
 
 
-DEFAULT_COLOR = Color("Grey", "#a9a9a9", True, 1)
+DEFAULT_COLOR = Color("Czarna", "#000000", True, 1)
 
 COLORS = {
-    "Red":      Color("Red", "#e6194B", True, 3),
-    "Green":    Color("Green", "#3cb44b", True, 3),
-    "Yellow":   Color("Yellow", "#ffe119", False, 1),
-    "Blue":     Color("Blue", "#4363d8", True, 1),
-    "Orange":   Color("Orange", "#f58231", True, 2),
-    "Purple":   Color("Purple", "#911eb4", True, 4),
-    "Cyan":     Color("Cyan", "#42d4f4", False, 3),
-    "Magenta":  Color("Magenta", "#f032e6", True, 3),
-    "Lime":     Color("Lime", "#bfef45", False, 4),
-    "Pink":     Color("Pink", "#fabed4", False, 3),
-    "Teal":     Color("Teal", "#469990", True, 3),
-    "Lavender": Color("Lavender", "#dcbeff", False, 2),
-    "Brown":    Color("Brown", "#9A6324", True, 3),
-    "Beige":    Color("Beige", "#fffac8", False, 3),
-    "Maroon":   Color("Maroon", "#800000", True, 2),
-    "Mint":     Color("Mint", "#aaffc3", False, 3),
-    "Olive":    Color("Olive", "#808000", True, 4),
-    "Apricot":  Color("Apricot", "#ffd8b1", False, 4),
-    "Navy":     Color("Navy", "#000075", True, 2),
-    "White":    Color("White", "#ffffff", False, 1),
-    "Black":    Color("Black", "#000000", True, 1)
+    "Czerwona":     Color("Czerwona", "#cd0032", True, 3),
+    "Zielona":      Color("Zielona", "#00700e", True, 3),
+    "Niebieska":    Color("Niebieska", "#3254d2", True, 1),
+    "Fioletowa":    Color("Fioletowa", "#911eb4", True, 4),
+    "Turkusowa":    Color("Turkusowa", "#156f82", False, 3),
+    "Magenta":      Color("Magenta", "#b900af", True, 3),
+    "Morska":       Color("Morska", "#077267", True, 3),
+    "Brązowa":      Color("Brązowa", "#975101", True, 3),
+    "Bordowa":      Color("Bordowa", "#800000", True, 2),
+    "Oliwkowa":     Color("Oliwkowa", "#696901", True, 4),
+    "Morelowa":     Color("Morelowa", "#ffd8b1", False, 4),
+    "Granatowa":    Color("Granatowa", "#000075", True, 2),
+    "Biała":        Color("Biała", "#ffffff", False, 1),
+    "Szara":        Color("Szara", "#a9a9a9", True, 1) 
 }
 
+# Rośliny VU, EN oraz CR na bazie https://pl.wikipedia.org/wiki/Polska_czerwona_ksi%C4%99ga_ro%C5%9Blin
 WORDS = [
-    "Aardvark",
-    "Alligator",
-    "Alpaca",
-    "Anaconda",
-    "Ant",
-    "Antelope",
-    "Ape",
-    "Aphid",
-    "Armadillo",
-    "Asp",
-    "Ass",
-    "Baboon",
-    "Badger",
-    "Bald Eagle",
-    "Barracuda",
-    "Bass",
-    "Basset Hound",
-    "Bat",
-    "Bear",
-    "Beaver",
-    "Bedbug",
-    "Bee",
-    "Beetle",
-    "Bird",
-    "Bison",
-    "Black panther",
-    "Black Widow Spider",
-    "Blue Jay",
-    "Blue Whale",
-    "Bobcat",
-    "Buffalo",
-    "Butterfly",
-    "Buzzard",
-    "Camel",
-    "Caribou",
-    "Carp",
-    "Cat",
-    "Caterpillar",
-    "Catfish",
-    "Cheetah",
-    "Chicken",
-    "Chimpanzee",
-    "Chipmunk",
-    "Cobra",
-    "Cod",
-    "Condor",
-    "Cougar",
-    "Cow",
-    "Coyote",
-    "Crab",
-    "Crane",
-    "Cricket",
-    "Crocodile",
-    "Crow",
-    "Cuckoo",
-    "Deer",
-    "Dinosaur",
-    "Dog",
-    "Dolphin",
-    "Donkey",
-    "Dove",
-    "Dragonfly",
-    "Duck",
-    "Eagle",
-    "Eel",
-    "Elephant",
-    "Emu",
-    "Falcon",
-    "Ferret",
-    "Finch",
-    "Fish",
-    "Flamingo",
-    "Flea",
-    "Fly",
-    "Fox",
-    "Frog",
-    "Goat",
-    "Goose",
-    "Gopher",
-    "Gorilla",
-    "Grasshopper",
-    "Hamster",
-    "Hare",
-    "Hawk",
-    "Hippopotamus",
-    "Horse",
-    "Hummingbird",
-    "Humpback Whale",
-    "Husky",
-    "Iguana",
-    "Impala",
-    "Kangaroo",
-    "Ladybug",
-    "Leopard",
-    "Lion",
-    "Lizard",
-    "Llama",
-    "Lobster",
-    "Mongoose",
-    "Monitor lizard",
-    "Monkey",
-    "Moose",
-    "Mosquito",
-    "Moth",
-    "Mountain goat",
-    "Mouse",
-    "Mule",
-    "Octopus",
-    "Orca",
-    "Ostrich",
-    "Otter",
-    "Owl",
-    "Ox",
-    "Oyster",
-    "Panda",
-    "Parrot",
-    "Peacock",
-    "Pelican",
-    "Penguin",
-    "Perch",
-    "Pheasant",
-    "Pig",
-    "Pigeon",
-    "Polar bear",
-    "Porcupine",
-    "Quail",
-    "Rabbit",
-    "Raccoon",
-    "Rat",
-    "Rattlesnake",
-    "Raven",
-    "Rooster",
-    "Sea lion",
-    "Sheep",
-    "Shrew",
-    "Skunk",
-    "Snail",
-    "Snake",
-    "Spider",
-    "Tiger",
-    "Walrus",
-    "Whale",
-    "Wolf",
-    "Zebra"
+    'Mokrzyca', 'Storczyk', 'Pierwiosnka', 'Ostnica', 'Dzwonek', 'Przygiełka', 'Saussurea', 'Jarząb', 'Miodokwiat', 'Kaldezja', 'Wełnianka', 'Babka', 'Czosnek', 'Turzyca', 'Lobelia', 'Szyplin', 'Gnidosz', 'Perz', 'Stoplamek', 'Ponikło', 'Ciemiężyca', 'Różanecznik', 'Irga', 'Cebulica', 'Łoboda', 'Wieczornik', 'Miłek', 'Storzan', 'Rozrzutka', 'Fiołek', 'Jeżogłówka', 'Karmnik', 'Kręczynka', 'Aldrowanda', 'Przetacznik', 'Stulisz', 'Głodek', 'Czechrzyca', 'Zmienka', 'Jaskier', 'Dziurawiec', 'Poryblin', 'Wierzba', 'Okrzyn', 'Elisma', 'Ostrołódka', 'Potrostek', 'Cyklamen', 'Starodub', 'Wiechlinostrzewa', 'Traganek', 'Dąb', 'Pszonacznik', 'Kiksja', 'Kruszczyk', 'Przesiąkra', 'Brzoza', 'Brzeżyca', 'Rzeżucha', 'Macierzanka', 'Gałuszka', 'Oman', 'Wątlik', 'Wełnianeczka', 'Żmijowiec', 'Rogownica', 'Groszek', 'Mniszek', 'Rozchodnik', 'Widlicz', 'Buławnik', 'Sierpik', 'Tojad', 'Sit', 'Ostrzew', 'Przewiercień', 'Mannica', 'Dyptam', 'Perłówka', 'Przymiotno', 'Lindernia', 'Języczka', 'Kropidło', 'Szachownica', 'Malina', 'Zaraza', 'Chamedafne', 'Mieczyk', 'Selery', 'Kotewka', 'Przytulia', 'Szafirek', 'Szczodrzeniec', 'Tłustosz', 'Skalnica', 'Zanokcica', 'Podejźrzon', 'Wawrzynek', 'Gęsiówka', 'Bylica', 'Wywłócznik', 'Nadbrzeżyca', 'Kukuczka', 'Rdestnica'
 ]
 
 Random(bytes('Nawiasem mówiąc: przedmioty są bezbarwne',
