+++ app/core/constants.py
@@ -2,5 +2,6 @@ DEBUG = True
 VERSION = '1.0-pre'
 
 ALLOW_DOWNLOAD = True
+FORCE_DOWNLOAD = False
 BRANCH = "prerelease"
 REPO_URL = f"https://raw.githubusercontent.com/Larch-Team/larch-plugins/{BRANCH}"
