+++ app/core/history.py
@@ -1,5 +1,6 @@
 from typing import Any, Callable, Union, Iterable
 from sentence import Sentence
+from constants import DEBUG
 def _f():
     pass
 
@@ -19,10 +20,13 @@ class History(set):
 
     def add_sentence(self, element: Sentence) -> None:
         """Dodaje zdanie `element` do zbioru, o ile już w nim nie jest. Konkretniej dodaje wartość hash zdania."""
-        if isinstance(element, Sentence):
-            return super().add(hash(element))
-        else:
+        if not isinstance(element, Sentence):
             raise TypeError("History can only store sentences")
+
+        if DEBUG:
+            return super().add(tuple(element.getUnique()))
+        else:
+            return super().add(hash(element))
 
     def __call__(self, *coms: tuple[Union[list, Sentence, int, Callable]]) -> None:
         """ Używane do manipulacji historią
@@ -55,6 +59,9 @@ class History(set):
         return super().__repr__()
 
     def __contains__(self, item: Sentence) -> bool:
+        if DEBUG:
+            return super().__contains__(tuple(item.getUnique()))
+        
         return super().__contains__(hash(item))
 
     def copy(self):
