+++ distributing/release-files/larch.cmd
@@ -1,0 +1,1 @@
+python larch.pyz
