+++ distributing/release-files/larch.sh
@@ -1,0 +1,1 @@
+python39 larch.pyz
